preventing infinite scroll
if the location is united states, ie when the remote jobs are being looked for, we can end up in a situation where it will try to find every job posted in the last day as per its design. Previously this was restricted enough by location that it was fine, but with having United States as the location, it makes it so that it is basically never ending.