Update Dockerfile
making this a better dockerfile. Grouping the run commands to better help with layers. Removed the need for cron by putting the schedule library into the python code itself. Made it so that the chrome driver that is downloaded is always the latest which will help ensure it works with the google chrome that is downloaded.